updated advanced

+++ 0x0C-more_malloc_free/101-mul.c
@@ -10,7 +10,7 @@
  */
 int main(int argc, char *argv[])
 {
-int num1, num2, sum, i, len1, len2;
+long int num1, num2, sum, i, len1, len2;
 if (argc != 3)
 {
 printf("Error\n");
@@ -37,6 +37,6 @@ exit(98);
 num1 = atoi(argv[1]);
 num2 = atoi(argv[2]);
 sum = num1 *num2;
-printf("%d\n", sum);
+printf("%ld\n", sum);
 return (0);
 }
